resumematch-backend/requirements.txt
# Core backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0
pymupdf==1.23.8

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
asyncpg==0.29.0

# Vector database
faiss-cpu==1.7.4
weaviate-client==4.6.5

# AI/NLP
transformers>=4.35.2
torch==2.1.1
tokenizers==0.15.0
sentence-transformers==2.2.2
spacy==3.7.2
openai==1.3.5

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
bcrypt==4.0.1
python-slugify==8.0.1
python-dateutil==2.8.2

# Async support
anyio==4.0.0
asyncio==3.4.3
concurrent-log-handler==0.9.20

# Environment
python-decouple==3.8
